requirements.txt
astroid==2.2.5
autopep8==1.4.4
dj-database-url==0.5.0
Django==2.2.3
django-cors-headers==3.0.2
django-filter==2.2.0
djangorestframework==3.9.4
gunicorn==19.9.0
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
Pillow==6.1.0
pycodestyle==2.5.0
pylint==2.3.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
whitenoise==4.1.3
wrapt==1.11.2
